:sparkles: Adds frail input argument and error output argument
This commit gives the user the ability to let the action fail when the
current version is not found. This can be done through the `frail` input
argument. When frail is set to 'false' still sets the `error` output
argumen to true to let the user know that an error occured.